+++ .php-cs-fixer.dist.php
@@ -2,7 +2,7 @@
 
 $finder = PhpCsFixer\Finder::create()
     ->in(__DIR__ . '/src')
-    ->in(__DIR__ . '/Tests')
+    ->in(__DIR__ . '/tests')
     ->name('*.php');
 
 return (new PhpCsFixer\Config())